Crate secretsharing_shamir

Shamir Secret Sharing (SSS)

This crate provides a complete randomized implementation of the Shamir Secret Sharing scheme in cryptography.

Overview

The scheme allows a secret to be split into multiple shares, such that only a threshold number of shares are required to reconstruct the original secret. Fewer than the threshold shares reveal nothing about it.

Core Features

Generate shares of a secret securely

Reconstruct the secret from a threshold of shares

Supports large prime sizes (256, 512, 1024 bits)

Polynomial degrees up to 255 (max size of u8) tested

Option to use fixed primes or generate them dynamically

Usage

Add this crate to your Cargo.toml:

secretsharing-shamir = "0.1"

Import the main types in your code:

use secretsharing_shamir::{SS, BitSize, Share, Error};
use num_bigint::{BigUint};

Initialize SS with four parameters:

SS {
    prime_size: BitSize,        // One of 256, 512, 1024
    use_fixed_prime: bool,      // Choose fixed primes or generate new
    threshold: u8,              // Minimum number of shares needed
    secret: &BigUint,           // The secret to be shared
}

For a complete working example, see examples/usage.rs.

Notes

Note 1: This implementation has been stress-tested with fixed and generated primes
of sizes 256, 512, and 1024 bits, with polynomials of degree up to 20.

Note 2: Available prime sizes: 256, 512, 1024 bits (choose fixed or generated).

Note 3: The reconstruction is tested to work for threhold or more no of parties and fails when threshold is not reached.

Structs

SS

SS struct for storing prime modulus(BigUint) and the polynomial()

Share

Share struct for storing (x,y) pairs where y = polynomial(x) mod prime

Enums

BitSize

BitSize enum for choosing bit sizes

Error

Enum for error handling